Add "classNameRowSelected" and "key" options
`classNameRowSelected` helps injecting css classes to the selected row.

Because the `svelte-table` renders a filtered-copy (`c_rows` variable) of the data source, adding an extended field to indicate the `selected` state to the copy is not a good idea because whenever the source data is changed, the copy will be updated and it will lose the extended field. Adding an extended field to the source data is not a good solution as well, because it cause an entire re-render to the table.
So adding `key` option to help SvelteTable keep track of row identity to do comparisons is a good idea, like the way Reactjs use `key` prop when rendering a list or Vuejs use `key` prop when using `v-for`.